Updated windows project files. Removed cout from ball constructor.

+++ Arc Engine/Arc Engine/Arc/IPhysicsObject.h
@@ -23,8 +23,8 @@ namespace Arc
 
         virtual void update( Vector2& pos, double delta )
         {
-            pos += Vel * delta;
-            Vel += Acc * delta;
+            pos += Vel * (float)delta;
+            Vel += Acc * (float)delta;
         }
 
     }; // class IPhysicsObject
